tidy up get_sns_batches
