+++ sierra_adapter/sierra_reader_new/sierra_reader.py
@@ -2,10 +2,10 @@ import datetime
 import json
 import os
 import sys
-import uuid
 
 import boto3
 
+from aws import get_sns_batches
 from client import catalogue_client
 
 
@@ -125,45 +125,6 @@ def to_scala_record(record):
     )
 
 
-def get_sns_batches(messages):
-    """
-    Given an iterable of SNS messages, group them into batches which
-    are suitable for sending to the SNS PublishBatch API.
-
-    There are two rules for batches:
-
-    *   You can send up to 10 messages in a single batch
-    *   A single batch can't exceed 256KB
-
-    """
-    this_batch = []
-
-    for m in messages:
-        this_message = [{"Id": str(uuid.uuid4()), "Message": m}]
-
-        old_batch = this_batch
-        new_batch = this_batch + this_message
-
-        # If adding this message to the batch will push us over the 256KB
-        # threshold, send all the other messages then start a new batch
-        # with the latest message.
-        if len(json.dumps(new_batch)) > 250 * 1024:
-            yield old_batch
-            this_batch = [this_message]
-            continue
-
-        # If there are 10 messages in this batch, send the entire batch
-        # and then start a new batch.
-        if len(new_batch) == 10:
-            yield new_batch
-            this_batch = []
-            continue
-
-    # If there are any leftover records, send them when we're done.
-    if this_batch:
-        yield this_batch
-
-
 def run_with(json_window):
     window = {
         k: datetime.datetime.fromisoformat(v) for k, v in json_window.items()
@@ -183,6 +144,7 @@ def run_with(json_window):
     )
 
     for batch_entries in get_sns_batches(sierra_records):
+        print(f"    Sending {len(batch_entries)} records to SNS…")
         resp = sns_client.publish_batch(
             TopicArn=topic_arn, PublishBatchRequestEntries=batch_entries
         )
